changed from in to double

+++ cs142/blankpracticesheet/ConsoleApplication1/blankpracticesheet.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 using namespace std;
-int main()
+double main()
 {
 	double a = 3;
 	double b = 4;
